apps/workload_starter/requirements.txt
ankaios-sdk==0.5.0rc4
paho-mqtt==2.1.0